requirements.txt
pythran-openblas
dvc
fastapi
imbalanced-learn
mlflow
nltk
numpy
optuna
pandas
rich
scikit-learn
streamlit
typer
uvicorn
inquirer
google-cloud-aiplatform
tensorflow-hub
<<<<<<< HEAD
tensorflow
=======
apache-airflow-providers-airbyte
>>>>>>> 43d79d0ee656231c282bcaaacc159ae0f8b1d569
